[gh-#672] working example for cli integration template
