+++ packages/cli/lib/index.ts
@@ -9,17 +9,13 @@ import axios from 'axios';
 import chalk from 'chalk';
 import figlet from 'figlet';
 import fs from 'fs';
-import path, { dirname } from 'path';
-import { fileURLToPath } from 'url';
-import yaml from 'js-yaml';
-import ejs from 'ejs';
+import path from 'path';
 import * as dotenv from 'dotenv';
 import promptly from 'promptly';
 
-import type { NangoConfig, NangoIntegration, NangoIntegrationData } from '@nangohq/shared';
 import { cloudHost, stagingHost, nangoConfigFile, loadSimplifiedConfig, checkForIntegrationFile } from '@nangohq/shared';
-import { init, run, tsc, tscWatch, configWatch, dockerRun } from './sync.js';
-import { hostport, checkEnvVars, enrichHeaders, httpsAgent, NANGO_INTEGRATIONS_LOCATION, buildInterfaces } from './utils.js';
+import { init, run, generate, tsc, tscWatch, configWatch, dockerRun } from './sync.js';
+import { checkEnvVars, enrichHeaders, httpsAgent, NANGO_INTEGRATIONS_LOCATION } from './utils.js';
 
 interface GlobalOptions {
     secretKey?: string;
@@ -37,9 +33,6 @@ class NangoCommand extends Command {
 }
 
 const program = new NangoCommand();
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
 
 dotenv.config();
 
@@ -65,52 +58,7 @@ program
     .alias('g')
     .description('Generate a new Nango integration')
     .action(() => {
-        const templateContents = fs.readFileSync(path.resolve(__dirname, './integration.ejs'), 'utf8');
-
-        const cwd = process.cwd();
-        const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
-        const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
-        const { integrations } = configData;
-        const { models } = configData;
-
-        const interfaceDefinitions = buildInterfaces(models);
-
-        fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaceDefinitions.join('\n'));
-
-        for (let i = 0; i < Object.keys(integrations).length; i++) {
-            const providerConfigKey = Object.keys(integrations)[i] as string;
-            const syncObject = integrations[providerConfigKey] as unknown as { [key: string]: NangoIntegration };
-            const syncNames = Object.keys(syncObject);
-            for (let k = 0; k < syncNames.length; k++) {
-                const syncName = syncNames[k] as string;
-                const syncData = syncObject[syncName] as unknown as NangoIntegrationData;
-                const { returns: models } = syncData;
-                const syncNameCamel = syncName
-                    .split('-')
-                    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-                    .join('');
-                const rendered = ejs.render(templateContents, {
-                    syncName: syncNameCamel,
-                    interfaceNames: models.map((model) => {
-                        const singularModel = model?.charAt(model.length - 1) === 's' ? model.slice(0, -1) : model;
-                        return `${singularModel.charAt(0).toUpperCase()}${singularModel.slice(1)}`;
-                    }),
-                    mappings: models.map((model) => {
-                        const singularModel = model.charAt(model.length - 1) === 's' ? model.slice(0, -1) : model;
-                        return {
-                            name: model,
-                            type: `${singularModel.charAt(0).toUpperCase()}${singularModel.slice(1)}`
-                        };
-                    })
-                });
-
-                if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/${syncName}.ts`)) {
-                    fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/${syncName}.ts`, rendered);
-                }
-            }
-        }
-
-        console.log(chalk.green(`Integration files have been created`));
+        generate();
     });
 
 program
@@ -190,7 +138,7 @@ program
                 }
             }
 
-            if (hostport !== 'http://localhost:3003' && !process.env['NANGO_SECRET_KEY']) {
+            if (process.env['NANGO_HOSTPORT'] !== 'http://localhost:3003' && !process.env['NANGO_SECRET_KEY']) {
                 console.log(chalk.red(`NANGO_SECRET_KEY environment variable is not set. Please set it now`));
                 try {
                     const secretKey = await promptly.prompt('Secret Key: ');
@@ -205,7 +153,7 @@ program
                 }
             }
 
-            checkEnvVars();
+            checkEnvVars(process.env['NANGO_HOSTPORT']);
             tsc();
 
             const cwd = process.cwd();
@@ -252,7 +200,7 @@ program
                 }
             }
 
-            const url = hostport + `/sync/deploy`;
+            const url = process.env['NANGO_HOSTPORT'] + `/sync/deploy`;
 
             await axios
                 .post(url, postData, { headers: enrichHeaders(), httpsAgent: httpsAgent() })
